Moved drag & drop plugin into its own component
